Finishing package API with rectangle and points geoms
I originally intended to have just a single geometry type, but it seems
that good balance is in fact to have two. These match the geom_point and
geom_rect layers from ggplot2. Specifying bounding boxes is best for full
control; using the points are nice for small icons or for EDA with image
datasets.